+++ CHANGELOG.md
@@ -5,6 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: `HeaderNav`: Add styles to bottom border, defaults to `'shadow'`.
+- Feature: `Modal/Next`: Add `fullSizeOnMobile` prop, defaults to false.
+- Feature: `Modal/Next`: Add the following props to `Modal.Actions`:
+  - `sticky`
+  - `stickyOnMobile`
+  - `fullSize`
+  - `fullSizeOnMobile`
+- Feature: `Modal/Next`: Deprecate `withoutMargin` prop on `Modal.Paragraph` in favor of `noMargin`.
 - Feature: Add `boron` modifier to Button component.
 - Feature: Add new color: `font3`.